Remove Windows class
